Refactor to split random & request based metrics
