+++ sample-apps/go-sample-app/collection/http_traces.go
